+++ ruby_runtime/android/src/main/kotlin/com/example/ruby_runtime/MrubyRuntimePlugin.kt
@@ -1,4 +1,4 @@
-package com.example.ruby_runtime
+package com.izeesoft.ruby_runtime
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall